HDA: Allow driver subclasing
