+++ sys/conf/files
@@ -3122,11 +3122,12 @@ dev/sound/pci/t4dwave.c		optional snd_t4dwave pci
 dev/sound/pci/via8233.c		optional snd_via8233 pci
 dev/sound/pci/via82c686.c	optional snd_via82c686 pci
 dev/sound/pci/vibes.c		optional snd_vibes pci
-dev/sound/pci/hda/hdaa.c	optional snd_hda pci
-dev/sound/pci/hda/hdaa_patches.c	optional snd_hda pci
-dev/sound/pci/hda/hdac.c	optional snd_hda pci
-dev/sound/pci/hda/hdac_if.m	optional snd_hda pci
-dev/sound/pci/hda/hdacc.c	optional snd_hda pci
+dev/sound/pci/hda/hdaa.c	optional snd_hda
+dev/sound/pci/hda/hdaa_patches.c	optional snd_hda
+dev/sound/pci/hda/hdac.c	optional snd_hda
+dev/sound/pci/hda/hdac_pci.c	optional snd_hda pci
+dev/sound/pci/hda/hdac_if.m	optional snd_hda
+dev/sound/pci/hda/hdacc.c	optional snd_hda
 dev/sound/pci/hdspe.c		optional snd_hdspe pci
 dev/sound/pci/hdspe-pcm.c	optional snd_hdspe pci
 dev/sound/pcm/ac97.c		optional sound
