+++ sys/dev/sound/pci/hda/hdac.c
@@ -37,8 +37,6 @@
 #endif
 
 #include <dev/sound/pcm/sound.h>
-#include <dev/pci/pcireg.h>
-#include <dev/pci/pcivar.h>
 
 #include <sys/ctype.h>
 #include <sys/endian.h>
@@ -57,10 +55,6 @@ SND_DECLARE_FILE("$FreeBSD$");
 #define hdac_unlock(sc)		snd_mtxunlock((sc)->lock)
 #define hdac_lockassert(sc)	snd_mtxassert((sc)->lock)
 
-#define HDAC_QUIRK_64BIT	(1 << 0)
-#define HDAC_QUIRK_DMAPOS	(1 << 1)
-#define HDAC_QUIRK_MSI		(1 << 2)
-
 static const struct {
 	const char *key;
 	uint32_t value;
@@ -72,147 +66,6 @@ static const struct {
 
 MALLOC_DEFINE(M_HDAC, "hdac", "HDA Controller");
 
-static const struct {
-	uint32_t	model;
-	const char	*desc;
-	char		quirks_on;
-	char		quirks_off;
-} hdac_devices[] = {
-	{ HDA_INTEL_OAK,     "Intel Oaktrail",	0, 0 },
-	{ HDA_INTEL_CMLKLP,  "Intel Comet Lake-LP",	0, 0 },
-	{ HDA_INTEL_CMLKH,   "Intel Comet Lake-H",	0, 0 },
-	{ HDA_INTEL_BAY,     "Intel BayTrail",	0, 0 },
-	{ HDA_INTEL_HSW1,    "Intel Haswell",	0, 0 },
-	{ HDA_INTEL_HSW2,    "Intel Haswell",	0, 0 },
-	{ HDA_INTEL_HSW3,    "Intel Haswell",	0, 0 },
-	{ HDA_INTEL_BDW1,    "Intel Broadwell",	0, 0 },
-	{ HDA_INTEL_BDW2,    "Intel Broadwell",	0, 0 },
-	{ HDA_INTEL_BXTNT,   "Intel Broxton-T",	0, 0 },
-	{ HDA_INTEL_CPT,     "Intel Cougar Point",	0, 0 },
-	{ HDA_INTEL_PATSBURG,"Intel Patsburg",  0, 0 },
-	{ HDA_INTEL_PPT1,    "Intel Panther Point",	0, 0 },
-	{ HDA_INTEL_BR,      "Intel Braswell",	0, 0 },
-	{ HDA_INTEL_LPT1,    "Intel Lynx Point",	0, 0 },
-	{ HDA_INTEL_LPT2,    "Intel Lynx Point",	0, 0 },
-	{ HDA_INTEL_WCPT,    "Intel Wildcat Point",	0, 0 },
-	{ HDA_INTEL_WELLS1,  "Intel Wellsburg",	0, 0 },
-	{ HDA_INTEL_WELLS2,  "Intel Wellsburg",	0, 0 },
-	{ HDA_INTEL_LPTLP1,  "Intel Lynx Point-LP",	0, 0 },
-	{ HDA_INTEL_LPTLP2,  "Intel Lynx Point-LP",	0, 0 },
-	{ HDA_INTEL_SRPTLP,  "Intel Sunrise Point-LP",	0, 0 },
-	{ HDA_INTEL_KBLKLP,  "Intel Kaby Lake-LP",	0, 0 },
-	{ HDA_INTEL_SRPT,    "Intel Sunrise Point",	0, 0 },
-	{ HDA_INTEL_KBLK,    "Intel Kaby Lake",	0, 0 },
-	{ HDA_INTEL_KBLKH,   "Intel Kaby Lake-H",	0, 0 },
-	{ HDA_INTEL_CFLK,    "Intel Coffee Lake",	0, 0 },
-	{ HDA_INTEL_CMLKS,   "Intel Comet Lake-S",	0, 0 },
-	{ HDA_INTEL_CNLK,    "Intel Cannon Lake",	0, 0 },
-	{ HDA_INTEL_ICLK,    "Intel Ice Lake",		0, 0 },
-	{ HDA_INTEL_CMLKLP,  "Intel Comet Lake-LP",	0, 0 },
-	{ HDA_INTEL_CMLKH,   "Intel Comet Lake-H",	0, 0 },
-	{ HDA_INTEL_TGLK,    "Intel Tiger Lake",	0, 0 },
-	{ HDA_INTEL_GMLK,    "Intel Gemini Lake",	0, 0 },
-	{ HDA_INTEL_82801F,  "Intel 82801F",	0, 0 },
-	{ HDA_INTEL_63XXESB, "Intel 631x/632xESB",	0, 0 },
-	{ HDA_INTEL_82801G,  "Intel 82801G",	0, 0 },
-	{ HDA_INTEL_82801H,  "Intel 82801H",	0, 0 },
-	{ HDA_INTEL_82801I,  "Intel 82801I",	0, 0 },
-	{ HDA_INTEL_JLK,     "Intel Jasper Lake",	0, 0 },
-	{ HDA_INTEL_82801JI, "Intel 82801JI",	0, 0 },
-	{ HDA_INTEL_82801JD, "Intel 82801JD",	0, 0 },
-	{ HDA_INTEL_PCH,     "Intel Ibex Peak",	0, 0 },
-	{ HDA_INTEL_PCH2,    "Intel Ibex Peak",	0, 0 },
-	{ HDA_INTEL_ELLK,    "Intel Elkhart Lake",	0, 0 },
-	{ HDA_INTEL_JLK2,    "Intel Jasper Lake",	0, 0 },
-	{ HDA_INTEL_BXTNP,   "Intel Broxton-P",	0, 0 },
-	{ HDA_INTEL_SCH,     "Intel SCH",	0, 0 },
-	{ HDA_NVIDIA_MCP51,  "NVIDIA MCP51",	0, HDAC_QUIRK_MSI },
-	{ HDA_NVIDIA_MCP55,  "NVIDIA MCP55",	0, HDAC_QUIRK_MSI },
-	{ HDA_NVIDIA_MCP61_1, "NVIDIA MCP61",	0, 0 },
-	{ HDA_NVIDIA_MCP61_2, "NVIDIA MCP61",	0, 0 },
-	{ HDA_NVIDIA_MCP65_1, "NVIDIA MCP65",	0, 0 },
-	{ HDA_NVIDIA_MCP65_2, "NVIDIA MCP65",	0, 0 },
-	{ HDA_NVIDIA_MCP67_1, "NVIDIA MCP67",	0, 0 },
-	{ HDA_NVIDIA_MCP67_2, "NVIDIA MCP67",	0, 0 },
-	{ HDA_NVIDIA_MCP73_1, "NVIDIA MCP73",	0, 0 },
-	{ HDA_NVIDIA_MCP73_2, "NVIDIA MCP73",	0, 0 },
-	{ HDA_NVIDIA_MCP78_1, "NVIDIA MCP78",	0, HDAC_QUIRK_64BIT },
-	{ HDA_NVIDIA_MCP78_2, "NVIDIA MCP78",	0, HDAC_QUIRK_64BIT },
-	{ HDA_NVIDIA_MCP78_3, "NVIDIA MCP78",	0, HDAC_QUIRK_64BIT },
-	{ HDA_NVIDIA_MCP78_4, "NVIDIA MCP78",	0, HDAC_QUIRK_64BIT },
-	{ HDA_NVIDIA_MCP79_1, "NVIDIA MCP79",	0, 0 },
-	{ HDA_NVIDIA_MCP79_2, "NVIDIA MCP79",	0, 0 },
-	{ HDA_NVIDIA_MCP79_3, "NVIDIA MCP79",	0, 0 },
-	{ HDA_NVIDIA_MCP79_4, "NVIDIA MCP79",	0, 0 },
-	{ HDA_NVIDIA_MCP89_1, "NVIDIA MCP89",	0, 0 },
-	{ HDA_NVIDIA_MCP89_2, "NVIDIA MCP89",	0, 0 },
-	{ HDA_NVIDIA_MCP89_3, "NVIDIA MCP89",	0, 0 },
-	{ HDA_NVIDIA_MCP89_4, "NVIDIA MCP89",	0, 0 },
-	{ HDA_NVIDIA_0BE2,   "NVIDIA (0x0be2)",	0, HDAC_QUIRK_MSI },
-	{ HDA_NVIDIA_0BE3,   "NVIDIA (0x0be3)",	0, HDAC_QUIRK_MSI },
-	{ HDA_NVIDIA_0BE4,   "NVIDIA (0x0be4)",	0, HDAC_QUIRK_MSI },
-	{ HDA_NVIDIA_GT100,  "NVIDIA GT100",	0, HDAC_QUIRK_MSI },
-	{ HDA_NVIDIA_GT104,  "NVIDIA GT104",	0, HDAC_QUIRK_MSI },
-	{ HDA_NVIDIA_GT106,  "NVIDIA GT106",	0, HDAC_QUIRK_MSI },
-	{ HDA_NVIDIA_GT108,  "NVIDIA GT108",	0, HDAC_QUIRK_MSI },
-	{ HDA_NVIDIA_GT116,  "NVIDIA GT116",	0, HDAC_QUIRK_MSI },
-	{ HDA_NVIDIA_GF119,  "NVIDIA GF119",	0, 0 },
-	{ HDA_NVIDIA_GF110_1, "NVIDIA GF110",	0, HDAC_QUIRK_MSI },
-	{ HDA_NVIDIA_GF110_2, "NVIDIA GF110",	0, HDAC_QUIRK_MSI },
-	{ HDA_ATI_SB450,     "ATI SB450",	0, 0 },
-	{ HDA_ATI_SB600,     "ATI SB600",	0, 0 },
-	{ HDA_ATI_RS600,     "ATI RS600",	0, 0 },
-	{ HDA_ATI_RS690,     "ATI RS690",	0, 0 },
-	{ HDA_ATI_RS780,     "ATI RS780",	0, 0 },
-	{ HDA_ATI_R600,      "ATI R600",	0, 0 },
-	{ HDA_ATI_RV610,     "ATI RV610",	0, 0 },
-	{ HDA_ATI_RV620,     "ATI RV620",	0, 0 },
-	{ HDA_ATI_RV630,     "ATI RV630",	0, 0 },
-	{ HDA_ATI_RV635,     "ATI RV635",	0, 0 },
-	{ HDA_ATI_RV710,     "ATI RV710",	0, 0 },
-	{ HDA_ATI_RV730,     "ATI RV730",	0, 0 },
-	{ HDA_ATI_RV740,     "ATI RV740",	0, 0 },
-	{ HDA_ATI_RV770,     "ATI RV770",	0, 0 },
-	{ HDA_ATI_RV810,     "ATI RV810",	0, 0 },
-	{ HDA_ATI_RV830,     "ATI RV830",	0, 0 },
-	{ HDA_ATI_RV840,     "ATI RV840",	0, 0 },
-	{ HDA_ATI_RV870,     "ATI RV870",	0, 0 },
-	{ HDA_ATI_RV910,     "ATI RV910",	0, 0 },
-	{ HDA_ATI_RV930,     "ATI RV930",	0, 0 },
-	{ HDA_ATI_RV940,     "ATI RV940",	0, 0 },
-	{ HDA_ATI_RV970,     "ATI RV970",	0, 0 },
-	{ HDA_ATI_R1000,     "ATI R1000",	0, 0 },
-	{ HDA_AMD_X370,      "AMD X370",	0, 0 },
-	{ HDA_AMD_X570,      "AMD X570",	0, 0 },
-	{ HDA_AMD_STONEY,    "AMD Stoney",	0, 0 },
-	{ HDA_AMD_RAVEN,     "AMD Raven",	0, 0 },
-	{ HDA_AMD_HUDSON2,   "AMD Hudson-2",	0, 0 },
-	{ HDA_RDC_M3010,     "RDC M3010",	0, 0 },
-	{ HDA_VIA_VT82XX,    "VIA VT8251/8237A",0, 0 },
-	{ HDA_SIS_966,       "SiS 966/968",	0, 0 },
-	{ HDA_ULI_M5461,     "ULI M5461",	0, 0 },
-	/* Unknown */
-	{ HDA_INTEL_ALL,  "Intel",		0, 0 },
-	{ HDA_NVIDIA_ALL, "NVIDIA",		0, 0 },
-	{ HDA_ATI_ALL,    "ATI",		0, 0 },
-	{ HDA_AMD_ALL,    "AMD",		0, 0 },
-	{ HDA_CREATIVE_ALL,    "Creative",	0, 0 },
-	{ HDA_VIA_ALL,    "VIA",		0, 0 },
-	{ HDA_SIS_ALL,    "SiS",		0, 0 },
-	{ HDA_ULI_ALL,    "ULI",		0, 0 },
-};
-
-static const struct {
-	uint16_t vendor;
-	uint8_t reg;
-	uint8_t mask;
-	uint8_t enable;
-} hdac_pcie_snoop[] = {
-	{  INTEL_VENDORID, 0x00, 0x00, 0x00 },
-	{    ATI_VENDORID, 0x42, 0xf8, 0x02 },
-	{    AMD_VENDORID, 0x42, 0xf8, 0x02 },
-	{ NVIDIA_VENDORID, 0x4e, 0xf0, 0x0f },
-};
 
 /****************************************************************************
  * Function prototypes
@@ -224,10 +77,6 @@ static void	hdac_dma_cb(void *, bus_dma_segment_t *, int, int);
 static int	hdac_dma_alloc(struct hdac_softc *,
 					struct hdac_dma *, bus_size_t);
 static void	hdac_dma_free(struct hdac_softc *, struct hdac_dma *);
-static int	hdac_mem_alloc(struct hdac_softc *);
-static void	hdac_mem_free(struct hdac_softc *);
-static int	hdac_irq_alloc(struct hdac_softc *);
-static void	hdac_irq_free(struct hdac_softc *);
 static void	hdac_corb_init(struct hdac_softc *);
 static void	hdac_rirb_init(struct hdac_softc *);
 static void	hdac_corb_start(struct hdac_softc *);
@@ -237,9 +86,8 @@ static void	hdac_attach2(void *);
 
 static uint32_t	hdac_send_command(struct hdac_softc *, nid_t, uint32_t);
 
-static int	hdac_probe(device_t);
-static int	hdac_attach(device_t);
-static int	hdac_detach(device_t);
+static int	hdac_irq_setup(device_t dev);
+static void	hdac_irq_teardown(device_t dev);
 static int	hdac_suspend(device_t);
 static int	hdac_resume(device_t);
 
@@ -687,115 +535,6 @@ hdac_dma_free(struct hdac_softc *sc, struct hdac_dma *dma)
 }
 
 /****************************************************************************
- * int hdac_mem_alloc(struct hdac_softc *)
- *
- * Allocate all the bus resources necessary to speak with the physical
- * controller.
- ****************************************************************************/
-static int
-hdac_mem_alloc(struct hdac_softc *sc)
-{
-	struct hdac_mem *mem;
-
-	mem = &sc->mem;
-	mem->mem_rid = PCIR_BAR(0);
-	mem->mem_res = bus_alloc_resource_any(sc->dev, SYS_RES_MEMORY,
-	    &mem->mem_rid, RF_ACTIVE);
-	if (mem->mem_res == NULL) {
-		device_printf(sc->dev,
-		    "%s: Unable to allocate memory resource\n", __func__);
-		return (ENOMEM);
-	}
-	mem->mem_tag = rman_get_bustag(mem->mem_res);
-	mem->mem_handle = rman_get_bushandle(mem->mem_res);
-
-	return (0);
-}
-
-/****************************************************************************
- * void hdac_mem_free(struct hdac_softc *)
- *
- * Free up resources previously allocated by hdac_mem_alloc.
- ****************************************************************************/
-static void
-hdac_mem_free(struct hdac_softc *sc)
-{
-	struct hdac_mem *mem;
-
-	mem = &sc->mem;
-	if (mem->mem_res != NULL)
-		bus_release_resource(sc->dev, SYS_RES_MEMORY, mem->mem_rid,
-		    mem->mem_res);
-	mem->mem_res = NULL;
-}
-
-/****************************************************************************
- * int hdac_irq_alloc(struct hdac_softc *)
- *
- * Allocate and setup the resources necessary for interrupt handling.
- ****************************************************************************/
-static int
-hdac_irq_alloc(struct hdac_softc *sc)
-{
-	struct hdac_irq *irq;
-	int result;
-
-	irq = &sc->irq;
-	irq->irq_rid = 0x0;
-
-	if ((sc->quirks_off & HDAC_QUIRK_MSI) == 0 &&
-	    (result = pci_msi_count(sc->dev)) == 1 &&
-	    pci_alloc_msi(sc->dev, &result) == 0)
-		irq->irq_rid = 0x1;
-
-	irq->irq_res = bus_alloc_resource_any(sc->dev, SYS_RES_IRQ,
-	    &irq->irq_rid, RF_SHAREABLE | RF_ACTIVE);
-	if (irq->irq_res == NULL) {
-		device_printf(sc->dev, "%s: Unable to allocate irq\n",
-		    __func__);
-		goto hdac_irq_alloc_fail;
-	}
-	result = bus_setup_intr(sc->dev, irq->irq_res, INTR_MPSAFE | INTR_TYPE_AV,
-	    NULL, hdac_intr_handler, sc, &irq->irq_handle);
-	if (result != 0) {
-		device_printf(sc->dev,
-		    "%s: Unable to setup interrupt handler (%d)\n",
-		    __func__, result);
-		goto hdac_irq_alloc_fail;
-	}
-
-	return (0);
-
-hdac_irq_alloc_fail:
-	hdac_irq_free(sc);
-
-	return (ENXIO);
-}
-
-/****************************************************************************
- * void hdac_irq_free(struct hdac_softc *)
- *
- * Free up resources previously allocated by hdac_irq_alloc.
- ****************************************************************************/
-static void
-hdac_irq_free(struct hdac_softc *sc)
-{
-	struct hdac_irq *irq;
-
-	irq = &sc->irq;
-	if (irq->irq_res != NULL && irq->irq_handle != NULL)
-		bus_teardown_intr(sc->dev, irq->irq_res, irq->irq_handle);
-	if (irq->irq_res != NULL)
-		bus_release_resource(sc->dev, SYS_RES_IRQ, irq->irq_rid,
-		    irq->irq_res);
-	if (irq->irq_rid == 0x1)
-		pci_release_msi(sc->dev);
-	irq->irq_handle = NULL;
-	irq->irq_res = NULL;
-	irq->irq_rid = 0x0;
-}
-
-/****************************************************************************
  * void hdac_corb_init(struct hdac_softc *)
  *
  * Initialize the corb registers for operations but do not start it up yet.
@@ -1048,59 +787,6 @@ hdac_send_command(struct hdac_softc *sc, nid_t cad, uint32_t verb)
 	return (sc->codecs[cad].response);
 }
 
-/****************************************************************************
- * Device Methods
- ****************************************************************************/
-
-/****************************************************************************
- * int hdac_probe(device_t)
- *
- * Probe for the presence of an hdac. If none is found, check for a generic
- * match using the subclass of the device.
- ****************************************************************************/
-static int
-hdac_probe(device_t dev)
-{
-	int i, result;
-	uint32_t model;
-	uint16_t class, subclass;
-	char desc[64];
-
-	model = (uint32_t)pci_get_device(dev) << 16;
-	model |= (uint32_t)pci_get_vendor(dev) & 0x0000ffff;
-	class = pci_get_class(dev);
-	subclass = pci_get_subclass(dev);
-
-	bzero(desc, sizeof(desc));
-	result = ENXIO;
-	for (i = 0; i < nitems(hdac_devices); i++) {
-		if (hdac_devices[i].model == model) {
-			strlcpy(desc, hdac_devices[i].desc, sizeof(desc));
-			result = BUS_PROBE_DEFAULT;
-			break;
-		}
-		if (HDA_DEV_MATCH(hdac_devices[i].model, model) &&
-		    class == PCIC_MULTIMEDIA &&
-		    subclass == PCIS_MULTIMEDIA_HDA) {
-			snprintf(desc, sizeof(desc), "%s (0x%04x)",
-			    hdac_devices[i].desc, pci_get_device(dev));
-			result = BUS_PROBE_GENERIC;
-			break;
-		}
-	}
-	if (result == ENXIO && class == PCIC_MULTIMEDIA &&
-	    subclass == PCIS_MULTIMEDIA_HDA) {
-		snprintf(desc, sizeof(desc), "Generic (0x%08x)", model);
-		result = BUS_PROBE_GENERIC;
-	}
-	if (result != ENXIO) {
-		strlcat(desc, " HDA Controller", sizeof(desc));
-		device_set_desc_copy(dev, desc);
-	}
-
-	return (result);
-}
-
 static void
 hdac_unsolq_task(void *context, int pending)
 {
@@ -1120,65 +806,20 @@ hdac_unsolq_task(void *context, int pending)
  * when this function is called. Setup everything that doesn't require
  * interrupts and defer probing of codecs until interrupts are enabled.
  ****************************************************************************/
-static int
-hdac_attach(device_t dev)
+int
+hdac_attach_subclass(device_t dev)
 {
 	struct hdac_softc *sc;
 	int result;
-	int i, devid = -1;
-	uint32_t model;
-	uint16_t class, subclass;
-	uint16_t vendor;
-	uint8_t v;
+	int i;
 
 	sc = device_get_softc(dev);
-	HDA_BOOTVERBOSE(
-		device_printf(dev, "PCI card vendor: 0x%04x, device: 0x%04x\n",
-		    pci_get_subvendor(dev), pci_get_subdevice(dev));
-		device_printf(dev, "HDA Driver Revision: %s\n",
-		    HDA_DRV_TEST_REV);
-	);
-
-	model = (uint32_t)pci_get_device(dev) << 16;
-	model |= (uint32_t)pci_get_vendor(dev) & 0x0000ffff;
-	class = pci_get_class(dev);
-	subclass = pci_get_subclass(dev);
-
-	for (i = 0; i < nitems(hdac_devices); i++) {
-		if (hdac_devices[i].model == model) {
-			devid = i;
-			break;
-		}
-		if (HDA_DEV_MATCH(hdac_devices[i].model, model) &&
-		    class == PCIC_MULTIMEDIA &&
-		    subclass == PCIS_MULTIMEDIA_HDA) {
-			devid = i;
-			break;
-		}
-	}
-
 	sc->lock = snd_mtxcreate(device_get_nameunit(dev), "HDA driver mutex");
 	sc->dev = dev;
 	TASK_INIT(&sc->unsolq_task, 0, hdac_unsolq_task, sc);
 	callout_init(&sc->poll_callout, 1);
 	for (i = 0; i < HDAC_CODEC_MAX; i++)
 		sc->codecs[i].dev = NULL;
-	if (devid >= 0) {
-		sc->quirks_on = hdac_devices[devid].quirks_on;
-		sc->quirks_off = hdac_devices[devid].quirks_off;
-	} else {
-		sc->quirks_on = 0;
-		sc->quirks_off = 0;
-	}
-	if (resource_int_value(device_get_name(dev),
-	    device_get_unit(dev), "msi", &i) == 0) {
-		if (i == 0)
-			sc->quirks_off |= HDAC_QUIRK_MSI;
-		else {
-			sc->quirks_on |= HDAC_QUIRK_MSI;
-			sc->quirks_off |= ~HDAC_QUIRK_MSI;
-		}
-	}
 	hdac_config_fetch(sc, &sc->quirks_on, &sc->quirks_off);
 	HDA_BOOTVERBOSE(
 		device_printf(sc->dev,
@@ -1191,82 +832,6 @@ hdac_attach(device_t dev)
 		sc->polling = 1;
 	else
 		sc->polling = 0;
-
-	pci_enable_busmaster(dev);
-
-	vendor = pci_get_vendor(dev);
-	if (vendor == INTEL_VENDORID) {
-		/* TCSEL -> TC0 */
-		v = pci_read_config(dev, 0x44, 1);
-		pci_write_config(dev, 0x44, v & 0xf8, 1);
-		HDA_BOOTHVERBOSE(
-			device_printf(dev, "TCSEL: 0x%02d -> 0x%02d\n", v,
-			    pci_read_config(dev, 0x44, 1));
-		);
-	}
-
-#if defined(__i386__) || defined(__amd64__)
-	sc->flags |= HDAC_F_DMA_NOCACHE;
-
-	if (resource_int_value(device_get_name(dev),
-	    device_get_unit(dev), "snoop", &i) == 0 && i != 0) {
-#else
-	sc->flags &= ~HDAC_F_DMA_NOCACHE;
-#endif
-		/*
-		 * Try to enable PCIe snoop to avoid messing around with
-		 * uncacheable DMA attribute. Since PCIe snoop register
-		 * config is pretty much vendor specific, there are no
-		 * general solutions on how to enable it, forcing us (even
-		 * Microsoft) to enable uncacheable or write combined DMA
-		 * by default.
-		 *
-		 * http://msdn2.microsoft.com/en-us/library/ms790324.aspx
-		 */
-		for (i = 0; i < nitems(hdac_pcie_snoop); i++) {
-			if (hdac_pcie_snoop[i].vendor != vendor)
-				continue;
-			sc->flags &= ~HDAC_F_DMA_NOCACHE;
-			if (hdac_pcie_snoop[i].reg == 0x00)
-				break;
-			v = pci_read_config(dev, hdac_pcie_snoop[i].reg, 1);
-			if ((v & hdac_pcie_snoop[i].enable) ==
-			    hdac_pcie_snoop[i].enable)
-				break;
-			v &= hdac_pcie_snoop[i].mask;
-			v |= hdac_pcie_snoop[i].enable;
-			pci_write_config(dev, hdac_pcie_snoop[i].reg, v, 1);
-			v = pci_read_config(dev, hdac_pcie_snoop[i].reg, 1);
-			if ((v & hdac_pcie_snoop[i].enable) !=
-			    hdac_pcie_snoop[i].enable) {
-				HDA_BOOTVERBOSE(
-					device_printf(dev,
-					    "WARNING: Failed to enable PCIe "
-					    "snoop!\n");
-				);
-#if defined(__i386__) || defined(__amd64__)
-				sc->flags |= HDAC_F_DMA_NOCACHE;
-#endif
-			}
-			break;
-		}
-#if defined(__i386__) || defined(__amd64__)
-	}
-#endif
-
-	HDA_BOOTHVERBOSE(
-		device_printf(dev, "DMA Coherency: %s / vendor=0x%04x\n",
-		    (sc->flags & HDAC_F_DMA_NOCACHE) ?
-		    "Uncacheable" : "PCIe snoop", vendor);
-	);
-
-	/* Allocate resources */
-	result = hdac_mem_alloc(sc);
-	if (result != 0)
-		goto hdac_attach_fail;
-	result = hdac_irq_alloc(sc);
-	if (result != 0)
-		goto hdac_attach_fail;
 
 	/* Get Capabilities */
 	result = hdac_get_capabilities(sc);
@@ -1340,6 +905,12 @@ hdac_attach(device_t dev)
 	hdac_corb_init(sc);
 	hdac_rirb_init(sc);
 
+	result = HDAC_IRQ_SETUP(sc->dev);
+	if (result != 0) {
+		device_printf(dev, "%s: Cannot setup interrupt (%x)\n",
+		     __func__, result);
+		goto hdac_attach_fail;
+	}
 	/* Defer remaining of initialization until interrupts are enabled */
 	sc->intrhook.ich_func = hdac_attach2;
 	sc->intrhook.ich_arg = (void *)sc;
@@ -1351,18 +922,18 @@ hdac_attach(device_t dev)
 	return (0);
 
 hdac_attach_fail:
-	hdac_irq_free(sc);
+	HDAC_IRQ_TEARDOWN(sc->dev);
 	if (sc->streams != NULL)
 		for (i = 0; i < sc->num_ss; i++)
 			hdac_dma_free(sc, &sc->streams[i].bdl);
 	free(sc->streams, M_HDAC);
 	hdac_dma_free(sc, &sc->rirb_dma);
 	hdac_dma_free(sc, &sc->corb_dma);
-	hdac_mem_free(sc);
 	snd_mtxfree(sc->lock);
 
 	return (ENXIO);
 }
+
 
 static int
 sysctl_hdac_pindump(SYSCTL_HANDLER_ARGS)
@@ -1719,8 +1290,8 @@ hdac_resume(device_t dev)
  *
  * Detach and free up resources utilized by the hdac device.
  ****************************************************************************/
-static int
-hdac_detach(device_t dev)
+int
+hdac_detach_subclass(device_t dev)
 {
 	struct hdac_softc *sc = device_get_softc(dev);
 	device_t *devlist;
@@ -1742,7 +1313,7 @@ hdac_detach(device_t dev)
 	hdac_reset(sc, false);
 	hdac_unlock(sc);
 	taskqueue_drain(taskqueue_thread, &sc->unsolq_task);
-	hdac_irq_free(sc);
+	HDAC_IRQ_TEARDOWN(dev);
 
 	for (i = 0; i < sc->num_ss; i++)
 		hdac_dma_free(sc, &sc->streams[i].bdl);
@@ -1754,7 +1325,6 @@ hdac_detach(device_t dev)
 		bus_dma_tag_destroy(sc->chan_dmat);
 		sc->chan_dmat = NULL;
 	}
-	hdac_mem_free(sc);
 	snd_mtxfree(sc->lock);
 	return (0);
 }
@@ -1801,7 +1371,7 @@ hdac_child_pnpinfo_str_method(device_t dev, device_t child, char *buf,
 	return (0);
 }
 
-static int
+int
 hdac_read_ivar(device_t dev, device_t child, int which, uintptr_t *result)
 {
 	struct hdac_softc *sc = device_get_softc(dev);
@@ -1823,17 +1393,15 @@ hdac_read_ivar(device_t dev, device_t child, int which, uintptr_t *result)
 	case HDA_IVAR_STEPPING_ID:
 		*result = sc->codecs[cad].stepping_id;
 		break;
-	case HDA_IVAR_SUBVENDOR_ID:
-		*result = pci_get_subvendor(dev);
-		break;
-	case HDA_IVAR_SUBDEVICE_ID:
-		*result = pci_get_subdevice(dev);
-		break;
 	case HDA_IVAR_DMA_NOCACHE:
 		*result = (sc->flags & HDAC_F_DMA_NOCACHE) != 0;
 		break;
 	case HDA_IVAR_STRIPES_MASK:
 		*result = (1 << (1 << sc->num_sdo)) - 1;
+		break;
+	case HDA_IVAR_SUBVENDOR_ID:
+	case HDA_IVAR_SUBDEVICE_ID:
+		*result = 0;
 		break;
 	default:
 		return (ENOENT);
@@ -2135,6 +1703,50 @@ hdac_unsol_free(device_t dev, device_t child, int tag)
 	hdac_poll_reinit(sc);
 }
 
+/****************************************************************************
+ * int hdac_irq_setup(struct hdac_softc *)
+ *
+ * Setup interrupt.
+ ****************************************************************************/
+static int
+hdac_irq_setup(device_t dev)
+{
+	struct hdac_softc *sc = device_get_softc(dev);
+	struct hdac_irq *irq;
+	int result;
+
+	irq = &sc->irq;
+	result = bus_setup_intr(sc->dev, irq->irq_res,
+	    INTR_MPSAFE | INTR_TYPE_AV, NULL, hdac_intr_handler, sc,
+	    &irq->irq_handle);
+	if (result != 0) {
+		device_printf(sc->dev,
+		    "%s: Unable to setup interrupt handler (%x)\n",
+		    __func__, result);
+		return (ENXIO);
+	}
+
+	return (0);
+}
+
+/****************************************************************************
+ * void hdac_irq_teardown(struct hdac_softc *)
+ *
+ * Teardown interrupt.
+ ****************************************************************************/
+static void
+hdac_irq_teardown(device_t dev)
+{
+	struct hdac_softc *sc = device_get_softc(dev);
+	struct hdac_irq *irq;
+
+	irq = &sc->irq;
+	if (irq->irq_res != NULL && irq->irq_handle != NULL) {
+		bus_teardown_intr(sc->dev, irq->irq_res, irq->irq_handle);
+		irq->irq_handle = NULL;
+	}
+}
+
 static uint8_t
 hdac_read_1(device_t dev, struct hdac_mem *mem, bus_size_t offs)
 {
@@ -2179,9 +1791,6 @@ hdac_write_4(device_t dev, struct hdac_mem *mem, bus_size_t offs, uint32_t val)
 
 static device_method_t hdac_methods[] = {
 	/* device interface */
-	DEVMETHOD(device_probe,		hdac_probe),
-	DEVMETHOD(device_attach,	hdac_attach),
-	DEVMETHOD(device_detach,	hdac_detach),
 	DEVMETHOD(device_suspend,	hdac_suspend),
 	DEVMETHOD(device_resume,	hdac_resume),
 	/* Bus interface */
@@ -2190,6 +1799,7 @@ static device_method_t hdac_methods[] = {
 	DEVMETHOD(bus_child_location_str, hdac_child_location_str),
 	DEVMETHOD(bus_child_pnpinfo_str, hdac_child_pnpinfo_str_method),
 	DEVMETHOD(bus_read_ivar,	hdac_read_ivar),
+	/* HDAC interface */
 	DEVMETHOD(hdac_get_mtx,		hdac_get_mtx),
 	DEVMETHOD(hdac_codec_command,	hdac_codec_command),
 	DEVMETHOD(hdac_stream_alloc,	hdac_stream_alloc),
@@ -2206,16 +1816,12 @@ static device_method_t hdac_methods[] = {
 	DEVMETHOD(hdac_write_1,		hdac_write_1),
 	DEVMETHOD(hdac_write_2,		hdac_write_2),
 	DEVMETHOD(hdac_write_4,		hdac_write_4),
+	DEVMETHOD(hdac_irq_setup,	hdac_irq_setup),
+	DEVMETHOD(hdac_irq_teardown,	hdac_irq_teardown),
 
 	DEVMETHOD_END
 };
 
-static driver_t hdac_driver = {
-	"hdac",
-	hdac_methods,
-	sizeof(struct hdac_softc),
-};
 
-static devclass_t hdac_devclass;
-
-DRIVER_MODULE(snd_hda, pci, hdac_driver, hdac_devclass, NULL, NULL);
+DEFINE_CLASS_0(hda_base, hdac_base_driver, hdac_methods,
+    sizeof(struct hdac_softc));
